core/utils/requirements.txt
termcolor
PyCrypto
